made changes in delete api so that only user who have added book can delete it

+++ routes/books.routes.js
@@ -71,8 +71,12 @@ booksRouter.post("/create", async(req, res) => {
 booksRouter.delete("/delete/:id", async(req, res) => {
     try {
         const role = req.body.role
+        const userId = req.body.userId
         const {id} = req.params
-        if (role === "CREATOR") {
+        let book = await BookModel.findOne({_id: id})
+        let userIdDB = book.userId
+
+        if (role === "CREATOR" && userIdDB === userId) {
             await BookModel.findByIdAndDelete(id);
             res.status(200).json({ "msg": 'Book deleted successfully' });
